I fixed the by_month and by_day and show controllers for showing posts so that they correctly show a 404 when you enter an invalid URL. It would be better if the 404 page was better (esp I think it should use google analytics).
Also, Site.full_url is terrible and awful and fucks up things like logging into the admin section while running on the local machine. It's really not necessary as far as I can tell. What is the use case here? I can't tell. It seems pointless. I took it out of the admin log in.

git-svn-id: http://shrub.ca:8080/svn/swc@72 a73b2e98-c212-4573-bb7c-f9b89b9be6a6