Deleted the upgrade script because I don't need/want that.
Eliminated MOST occurrences of Site.full_url. Why? Because it's totally unnecessary. However in a couple of places it's useful for permalinks and the like.

Added a simple display_404 function in application.rb, useful as a short hand when you need to 404.

Eliminated notfound 'cause I don't need it any more.

git-svn-id: http://shrub.ca:8080/svn/swc@73 a73b2e98-c212-4573-bb7c-f9b89b9be6a6