Converted tags system from acts_as_taggable (the really old version, not the gem one), which is not just deprecated but gone in Rails 2.2, to has_many_polymorphs, which seems to be the de-facto replacement.
Also commented out the XML-RPC controller system, because it's broken in Rails 2.2.2, and anyway, I don't use it. Restoring it might be a good idea...

I've also reduced the GPL notice in the source code files to just one line, because I don't like having so much crap at the top, and it's not really necessary. And I added a CONTRIBUTORS file.

git-svn-id: http://shrub.ca:8080/svn/swc@116 a73b2e98-c212-4573-bb7c-f9b89b9be6a6