Updating cap deploy settings to work with new server.
In order to work with capistrano, I'm moving the rails app up to the top level of the repository. I suppose this is the "standard" way to do things in git anyway.

There's a rebase going on here, which will hopefully remove a whole bunch of unnecessary binary files from my old site from the reposity. Rebase is from d248ba6 to squash umm... I don't have the number, the suck.